completed setting card properly and making snap scroll

+++ app/src/main/java/com/omada/junction/JunctionUtilities.java
@@ -7,8 +7,8 @@ public class JunctionUtilities {
 
     //static utility functions like number converters, etc go here
 
-    public static int DP_to_PX(Context context, int dp_value){
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp_value, context.getResources().getDisplayMetrics());
+    public static float DP_to_PX(Context context, float dp_value){
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp_value, context.getResources().getDisplayMetrics());
     }
 
 }
